move ktor stub generator under ktor dir
