+++ codegen/smithy-kotlin-codegen/src/main/kotlin/software/amazon/smithy/kotlin/codegen/service/ktor/ServerFrameworkImplementation.kt
@@ -5,7 +5,6 @@ import software.amazon.smithy.kotlin.codegen.core.RuntimeTypes
 import software.amazon.smithy.kotlin.codegen.core.withBlock
 import software.amazon.smithy.kotlin.codegen.core.withInlineBlock
 import software.amazon.smithy.kotlin.codegen.lang.KotlinTypes
-import software.amazon.smithy.kotlin.codegen.service.KtorStubGenerator
 import software.amazon.smithy.kotlin.codegen.service.ServiceTypes
 
 internal fun KtorStubGenerator.writeServerFrameworkImplementation(writer: KotlinWriter) {
