+++ codegen/smithy-kotlin-codegen/src/main/kotlin/software/amazon/smithy/kotlin/codegen/service/ktor/Utils.kt
@@ -2,8 +2,6 @@ package software.amazon.smithy.kotlin.codegen.service.ktor
 
 import software.amazon.smithy.kotlin.codegen.core.RuntimeTypes
 import software.amazon.smithy.kotlin.codegen.core.withBlock
-import software.amazon.smithy.kotlin.codegen.service.KtorStubGenerator
-import software.amazon.smithy.kotlin.codegen.service.LoggingWriter
 import software.amazon.smithy.kotlin.codegen.service.ServiceTypes
 
 internal fun KtorStubGenerator.writeUtils() {
